# Conflicts:
#	src/GlobaleInteraction/GlobaleInteraction.js
#	src/GlobaleInteraction/Grid/Grid.js
#	src/GlobaleInteraction/Module/CellFactory/CellFactory.js
#	src/style.scss
#	src/utils/Canvas3D/Canvas3D.js

+++ src/GlobaleInteraction/Grid/gridMouseUp/gridMouseUp.js
@@ -1,22 +1,24 @@
 import gridDrag from "../gridDrag";
 export default function() {
   return e => {
-    window.removeEventListener("mousemove", this.dragEvent);
-
-    typeof this.pointer ? this.pointer.delete() : null;
-    if (this.sliceType == "horizontal") {
-      if (Math.abs(this.point.y - this.point2.y) > this.height) {
-        this.sliceHorizontal({
-          cuttingPoint: this.point.x,
-          direction: e.clientX > this.originSlice.x ? "right" : "left"
-        });
-      }
-    } else if (this.sliceType == "vertical") {
-      if (Math.abs(this.point.x - this.point2.x) > this.width) {
-        this.sliceVertical({
-          cuttingPoint: this.point.y,
-          direction: e.clientY > this.originSlice.xy ? "top" : "bottom"
-        });
+    if (this.isEventCreated) {
+      window.removeEventListener("mousemove", this.dragEvent);
+      this.isEventCreated = false;
+      typeof this.pointer ? this.pointer.delete() : null;
+      if (this.sliceType == "horizontal") {
+        if (Math.abs(this.point.y - this.point2.y) > this.height) {
+          this.sliceHorizontal({
+            cuttingPoint: this.point.x,
+            direction: e.clientX > this.originSlice.x ? "right" : "left"
+          });
+        }
+      } else if (this.sliceType == "vertical") {
+        if (Math.abs(this.point.x - this.point2.x) > this.width) {
+          this.sliceVertical({
+            cuttingPoint: this.point.y,
+            direction: e.clientY > this.originSlice.xy ? "top" : "bottom"
+          });
+        }
       }
     }
   };
